+++ dart/test/http_client/sentry_http_client_test.dart
@@ -28,12 +28,11 @@ void main() {
       final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'GET',
-        'status_code': 200,
-        'reason': 'OK',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['reason'], 'OK');
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('GET: happy path for 404', () async {
@@ -53,12 +52,11 @@ void main() {
       final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'GET',
-        'status_code': 404,
-        'reason': 'NOT FOUND',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['status_code'], 404);
+      expect(breadcrumb.data?['reason'], 'NOT FOUND');
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('POST: happy path', () async {
@@ -78,11 +76,10 @@ void main() {
       final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'POST',
-        'status_code': 200,
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'POST');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('PUT: happy path', () async {
@@ -102,11 +99,11 @@ void main() {
       final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'PUT',
-        'status_code': 200,
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'PUT');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['reason'], 'NOT FOUND');
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('DELETE: happy path', () async {
@@ -126,19 +123,16 @@ void main() {
       final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'DELETE',
-        'status_code': 200,
-        'reason': 'OK',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'DELETE');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     /// Tests, that in case an exception gets thrown, that
-    ///   - no breadcrumb gets added
-    ///   - no exception gets reported by Sentry, in case the user wants to
-    ///     handle the exception
-    test('no breadcrumb for ClientException', () async {
+    /// no exception gets reported by Sentry, in case the user wants to
+    /// handle the exception
+    test('no captureException for ClientException', () async {
       final url = Uri.parse('https://example.com');
 
       final mockHub = MockHub();
@@ -158,13 +152,12 @@ void main() {
         expect(e.uri, url);
       }
 
-      expect(mockHub.addBreadcrumbCalls.length, 0);
       expect(mockHub.captureExceptionCalls.length, 0);
     });
 
     /// SocketException are only a thing on dart:io platforms.
     /// otherwise this is equal to the test above
-    test('no breadcrumb for SocketException', () async {
+    test('no captureException for SocketException', () async {
       final url = Uri.parse('https://example.com');
 
       final mockHub = MockHub();
@@ -183,8 +176,35 @@ void main() {
         expect(e.message, 'test');
       }
 
-      expect(mockHub.addBreadcrumbCalls.length, 0);
       expect(mockHub.captureExceptionCalls.length, 0);
+    });
+
+    test('breadcrumb gets added when an exception gets thrown', () async {
+      final url = Uri.parse('https://example.com');
+
+      final mockHub = MockHub();
+
+      final mockClient = MockClient((request) async {
+        expect(request.url, url);
+        throw Exception('foo bar');
+      });
+
+      final client = SentryHttpClient(client: mockClient, hub: mockHub);
+
+      try {
+        await client.get(Uri.parse('https://example.com'));
+        fail('Method did not throw');
+      } on Exception catch (_) {}
+
+      expect(mockHub.addBreadcrumbCalls.length, 1);
+
+      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+
+      expect(breadcrumb.type, 'http');
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.level, SentryLevel.error);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('close does get called for user defined client', () async {
@@ -198,6 +218,28 @@ void main() {
       expect(mockHub.addBreadcrumbCalls.length, 0);
       expect(mockHub.captureExceptionCalls.length, 0);
       verify(mockClient.close());
+    });
+
+    test('Breadcrumb has correct duration', () async {
+      final mockHub = MockHub();
+
+      final mockClient = MockClient((request) async {
+        expect(request.url, Uri.parse('https://example.com'));
+        await Future.delayed(Duration(seconds: 1));
+        return Response('', 200, reasonPhrase: 'OK');
+      });
+
+      final client = SentryHttpClient(client: mockClient, hub: mockHub);
+
+      final response = await client.get(Uri.parse('https://example.com'));
+      expect(response.statusCode, 200);
+
+      expect(mockHub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+
+      var durationString = breadcrumb.data!['duration']! as String;
+      // we don't check for anything below a second
+      expect(durationString.startsWith('0:00:01'), true);
     });
   });
 }