+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix: `Sentry.close()` closes native SDK integrations (#388)
 * Fix: Mark `Sentry.currentHub` as deprecated (#406)
+* Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests
 
 # 5.0.0